Fix the np.float depreciation issue

+++ dppy/exact_sampling.py
@@ -1282,7 +1282,7 @@ def k_dpp_eig_vecs_selector(eig_vals, eig_vecs, size,
     N, k = eig_vecs.shape[0], size
 
     # as in np.linalg.matrix_rank
-    tol = np.max(eig_vals) * N * np.finfo(np.float).eps
+    tol = np.max(eig_vals) * N * np.finfo(float).eps
     rank = np.count_nonzero(eig_vals > tol)
     if k > rank:
         raise ValueError('size k={} > rank={}'.format(k, rank))
